openlink_deal_info {FIRVrBOE}	R Documentation

Get deal information for a vector of deal numbers

Description

Get deal information for a vector of deal numbers

Usage

openlink_deal_info(OL_deal_number)

Arguments

OL_deal_number
A vector of Openlink deal numbers

Value

A data_frame of information or a zero row data_frame in case of an error

Examples

openlink_deal_info(5089359)

[Package FIRVrBOE version 0.0.0.9000 Index]