openlink_query_vector_to_api_string {FIRVrBOE}	R Documentation

Convert a vector to the format required for Openlink API parameters

Description

Convert a vector to the format required for Openlink API parameters

Usage

openlink_query_vector_to_api_string(v)

Arguments

v
A vector of values

Value

A string of quoted, comma-separated values

Examples

openlink_query_vector_to_api_string(c("GBND", "IRS"))

[Package FIRVrBOE version 0.0.0.9000 Index]